MSEARCH-97: Split SystemUserService into in-memory and DB implementations.
